Refactor input validation
