+++ includes/inputvalidation.php
@@ -1,0 +1,11 @@
+<?php
+
+function validate($value) {
+    $value = trim($value);
+    $value = stripslashes($value);
+    $value = htmlspecialchars($value);
+    $value = htmlentities($value);
+    return $value;
+}
+
+?>
